+++ phoenix/wizard/views/done.py
@@ -68,7 +68,7 @@ class Done(Wizard):
             selection = self.wizard_state.get('wizard_esgf_search')['selection']
             source = json.loads(selection)
             source['url'] = self.request.registry.settings.get('esgfsearch.url')
-            source['credentials'] = user.get('credentials')
+            # source['credentials'] = user.get('credentials')
             workflow['source']['esgf'] = source
         elif 'solr' in source_type:
             state = self.wizard_state.get('wizard_solr')
@@ -91,12 +91,9 @@ class Done(Wizard):
         literal_inputs = appstruct_to_inputs(
             self.request,
             self.wizard_state.get('wizard_literal_inputs', {}))
-        # worker_inputs = ['%s=%s' % (key, value) for key, value in inputs]
-        # Use real wps url ... not proxy url.
-        service = self.request.catalog.get_service_by_name(self.service_name)
+        # Use proxy wps url ...
         wps = WebProcessingService(
-            #url=self.request.route_url('owsproxy', service_name=self.service_name),
-            url=service['url'],
+            url=self.request.route_url('owsproxy', service_name=self.service_name),
             verify=False, skip_caps=True)
         logger.debug("wizard worker wps url: %s", wps.url)
         worker = dict(